moved LinePasser to another file. Added library
