+++ Voxelzator/Voxelzator.sln
@@ -2,6 +2,8 @@
 Microsoft Visual Studio Solution File, Format Version 12.00
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Voxelzator", "Voxelzator\Voxelzator.csproj", "{76400B54-07A0-4B10-8BF8-1E6400C5483B}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Library", "Library\Library.csproj", "{56680DD0-E7CC-43DB-A037-1FF91BAB68E9}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -12,5 +14,9 @@ Global
 		{76400B54-07A0-4B10-8BF8-1E6400C5483B}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{76400B54-07A0-4B10-8BF8-1E6400C5483B}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{76400B54-07A0-4B10-8BF8-1E6400C5483B}.Release|Any CPU.Build.0 = Release|Any CPU
+		{56680DD0-E7CC-43DB-A037-1FF91BAB68E9}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{56680DD0-E7CC-43DB-A037-1FF91BAB68E9}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{56680DD0-E7CC-43DB-A037-1FF91BAB68E9}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{56680DD0-E7CC-43DB-A037-1FF91BAB68E9}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 EndGlobal
