+++ Voxelzator/Voxelzator/Program.cs
@@ -7,8 +7,8 @@ namespace Voxelzator
         static void Main(string[] args)
         {
             Solver solver = new Solver();
-            Point[] triangle = new Point[] { new Point(0, 0), new Point(1, 5), new Point(3, 4)};
-            var list = solver.CreateListOfRectangles(triangle, 1);
+            Point[] triangle = new Point[] { new Point(1, 1), new Point(1, 6), new Point(4, 1)};
+            var list = solver.CreateListOfRectangles(triangle, 0.5);
             Console.WriteLine(list.Count);
             
             foreach (Point cube in list)
